requirements.txt
gdown>=4.7.1
ultralytics>=8.0.0
Pillow>=9.0.0
streamlit>=1.20.0
langchain>=0.0.239
openai>=0.27.0
chromadb>=0.3.25
langchain-community>=0.3.9
requests>=2.32.3
tiktoken>=0.8.0